requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-google==10.3.0 # install airflow for google cloud (we dont need this one anymore after having python environment that already include soda and apache-airflow.
soda-core-bigquery==3.1.5   #install soda for google cloud

astronomer-cosmos [dbt-bigquery]==1.3.2
protobuf==4.25.2